Removed PSDQ directory. Updated paths.
