+++ src/views/Apply/Forms/Parts/CharacterChecks.vue
